:art: apply recommendations from PR review
