+++ elements/reigns/src/features/game/validateGameDefinition.test.ts
@@ -4,7 +4,6 @@ import {
   getFlags,
   validateFlags,
   getConditions,
-  validateAssetsUrl,
 } from "./validateGameDefinition";
 import gdpr from "../../../public/games/gdpr.json";
 import dont_starve from "../../../public/games/dont-starve.json";
@@ -136,61 +135,6 @@ describe("validateGameDefinition", () => {
       const iterator = dont_starve.cards.values();
       const cards: Card[] = Array.from(iterator);
       expect(() => validateCards(cards)).not.toThrow();
-    });
-  });
-
-  describe("validateAssetsUrl", () => {
-    it("returns a default empty string", () => {
-      const foo = null as unknown as string;
-      expect(validateAssetsUrl(foo)).toBe("");
-    });
-
-    it("keep a default https url ", () => {
-      expect(validateAssetsUrl("https://foo.fr/mar/fi")).toBe(
-        "https://foo.fr/mar/fi"
-      );
-    });
-
-    it("remove a trailing slash for an https url ", () => {
-      expect(validateAssetsUrl("https://foo.fr/mar/fi/knine/")).toBe(
-        "https://foo.fr/mar/fi/knine"
-      );
-    });
-
-    it("uses the location pathname to prefix non http url", () => {
-      global.document = {
-        location: {
-          pathname: "/",
-        },
-      } as any;
-      expect(validateAssetsUrl("foo")).toBe("/foo");
-    });
-
-    it("when pathaname is a file, uses the current pathname folder to prefix non http url", () => {
-      global.document = {
-        location: {
-          pathname: "/foo/bar/baz",
-        },
-      } as any;
-      expect(validateAssetsUrl("qux")).toBe("/foo/bar/qux");
-    });
-
-    it("when pathname is a folder, uses the location pathname to prefix non http url", () => {
-      global.document = {
-        location: {
-          pathname: "/foo/bar/baz/",
-        },
-      } as any;
-      expect(validateAssetsUrl("qux")).toBe("/foo/bar/baz/qux");
-    });
-
-    it("remove trailing slash from provided url", () => {
-      global.document = {
-        location: {
-          pathname: "/foo/e",
-        },
-      } as any;
-      expect(validateAssetsUrl("quzz/")).toBe("/foo/quzz");
     });
   });
 });
