+++ elements/reigns/src/features/game/validateGameDefinition.ts
@@ -1,14 +1,15 @@
+import { getRootAssetsUrl } from "./gameDefinitionUtils";
 import { Card, CardFlag, GameDefinition } from "./types";
 
 export const validateGameDefinition = (
   definition: GameDefinition
 ): GameDefinition => ({
   ...definition,
-  assetsUrl: validateAssetsUrl(definition.assetsUrl),
+  assetsUrl: getRootAssetsUrl(definition.assetsUrl),
   ...validateCards(definition.cards),
 });
 
-const urlWithoutTrailingSlash = (url: string) => {
+export const urlWithoutTrailingSlash = (url: string) => {
   if (!url) return "";
 
   if (url.slice(-1) !== "/") {
@@ -16,35 +17,6 @@ const urlWithoutTrailingSlash = (url: string) => {
   }
 
   return url.slice(0, -1);
-};
-
-export const validateAssetsUrl = (url: string) => {
-  debugger;
-  if (!url) return "";
-
-  if (
-    url.startsWith("http://") ||
-    url.startsWith("https://") ||
-    url.startsWith("/")
-  ) {
-    return urlWithoutTrailingSlash(url);
-  }
-
-  const isFolder = document.location.pathname.slice(-1) === "/";
-  if (isFolder) {
-    return urlWithoutTrailingSlash(
-      `${urlWithoutTrailingSlash(document.location.pathname)}/${url}`
-    );
-  }
-
-  const [_path, ...reversedFolders] = document.location.pathname
-    .split("/")
-    .reverse();
-
-  debugger;
-  return urlWithoutTrailingSlash(
-    `${reversedFolders.reverse().join("/")}/${url}`
-  );
 };
 
 export const validateCards = (cards: Card[] | undefined) => {
